WIP - refactor server package
