+++ internal/server/router.go
@@ -3,8 +3,8 @@ package server
 import (
 	"net/http"
 
+	"github.com/ad9311/hitomgr/internal/controller"
 	"github.com/go-chi/chi"
-	"github.com/go-chi/chi/middleware"
 )
 
 const (
@@ -18,21 +18,21 @@ const (
 func routes() http.Handler {
 	mux := chi.NewRouter()
 
-	mux.Use(middleware.Recoverer)
+	// mux.Use(middleware.Recoverer)
 	mux.Use(loadSession)
 	mux.Use(newCSRF)
 
-	mux.Get(root, getRoot)
+	// mux.Get(root, getRoot)
 
-	mux.Get(dashboard, getDashboard)
+	// mux.Get(dashboard, getDashboard)
 
-	mux.Get(signIn, getSignIn)
-	mux.Post(signIn, postSignIn)
+	mux.Get(signIn, controller.GetSignIn)
+	// mux.Post(signIn, postSignIn)
 
-	mux.Get(signUp, getSignUp)
-	mux.Post(signUp, postSignUp)
+	// mux.Get(signUp, getSignUp)
+	// mux.Post(signUp, postSignUp)
 
-	mux.Post(signOut, postSignOut)
+	// mux.Post(signOut, postSignOut)
 
 	fileServer := http.FileServer(http.Dir("./web/static/"))
 	mux.Handle("/static/*", http.StripPrefix("/static/", fileServer))
